Example Database without any data set.
I'll add graphic schema of this db later but this is the one that I used for this project. Remember this has no data in it right now. Simply import this to your server and you can use it easily. You can use program it self to add data. But there's some tables that requires inputting by hand. So remember to do them.